more input checking
